routing/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
python-dotenv==1.0.0
pydantic==1.10.13
google-generativeai==0.3.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.2
